Switch to using size_t where needed

+++ src/omake/Eval.cpp
@@ -197,10 +197,10 @@ void Eval::StripLeadingSpaces(std::string& value)
     if (n != 0)
         value.replace(0, n, "");
 }
-size_t Eval::MacroSpan(const std::string line, int pos)
+size_t Eval::MacroSpan(const std::string line, size_t pos)
 {
     std::list<char> stack;
-    int pos1 = pos;
+    size_t pos1 = pos;
     while (line[pos1] == '$')
         pos1++;
     if (line[pos1] != '(' && line[pos1] != '{')
@@ -243,7 +243,7 @@ std::string Eval::ParseMacroLine(const std::string& in)
             {
                 rv += "$(";
                 std::string temp = in.substr(n + 2, m - 3);
-                for (int q = 0; q < temp.size() - 1; q++)
+                for (size_t q = 0; q < temp.size() - 1; q++)
                 {
                     if (temp[q] == '$')
                     {
@@ -692,11 +692,11 @@ std::string Eval::ReplaceStem(const std::string& stem, const std::string& patter
 size_t Eval::MatchesPattern(const std::string& name, const std::string& pattern, size_t& start, size_t begin)
 {
     size_t rv = std::string::npos;
-    int m = FindPercent(pattern, begin);
+    size_t m = FindPercent(pattern, begin);
     if (m != std::string::npos)
     {
-        int m1 = m;
-        int m2 = m;
+        size_t m1 = m;
+        size_t m2 = m;
         while (m1 && pattern[m1 - 1] != ' ')
             m1--;
         while (m2 < pattern.size() && pattern[m2] != ' ')
@@ -785,7 +785,7 @@ bool Eval::ThreeArgs(const std::string& line, std::string& one, std::string& two
 int Eval::GetNumber(const std::string& line)
 {
     int rv = 0;
-    int i;
+    size_t i;
     size_t n = line.find_first_not_of(' ');
     size_t m = line.find_last_not_of(' ');
     if (m == std::string::npos)
